+++ csvkitcat/alltext.py
@@ -11,7 +11,7 @@ MyIO = namedtuple('MyIO', ['rows', 'column_names', 'column_ids', 'output',])
 
 class AllTextUtility(CSVKitUtility):
 
-    def init_io(self):
+    def init_io(self, write_header=True):
         self.args.sniff_limit = 0       # TK is needed??
         self.args.no_inference = True   # TK is needed??
 
@@ -21,7 +21,8 @@ class AllTextUtility(CSVKitUtility):
         rows, column_names, column_ids = self.get_rows_and_column_names_and_column_ids(**self.reader_kwargs)
 
         output = agate.csv.writer(self.output_file, **self.writer_kwargs)
-        output.writerow(column_names)
+        if write_header is True:
+            output.writerow(column_names)
 
         return MyIO(rows=rows, column_names=column_names, column_ids=column_ids, output=output)